Changed manual-tests to  network-tests. Stripped out unneeded portions of hack/docker for one node network. Add make network-test. Added building needed wasm with docker buildenv.
