--- hack/docker/scripts/lrt/Pipfile
@@ -1,14 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-dramatiq = {extras = ["rabbitmq", "watch"],version = "*"}
-casperlabs-client = "*"
-jsonpickle = "*"
-
-[requires]
-python_version = "3.6"
