+++ hack/key-management/docker-gen-keys.sh
@@ -20,27 +20,13 @@ fi
 
 shift
 
-if [[ -z "$DRONE_BUILD_NUMBER" ]]; then
-    TAG=${CL_VERSION:-"latest"}
-    # Only pull if CL_VERSION isn't stated.  This allows using new version before published.
-    if [[ "$CL_VERSION" != "latest" ]]; then
-      docker pull casperlabs/key-generator:"$TAG" &> /dev/null || {
-          TAG="dev"
-          #echo "Failed to pull casperlabs/key-generator:latest"
-          #echo "Falling back to casperlabs/key-generator:dev"
-          docker pull casperlabs/key-generator:"$TAG" &> /dev/null
-      }
-    fi
-    docker run --rm -it --user $(id -u):$(id -g) -v "$OUTPUT_DIR":/keys casperlabs/key-generator:"$TAG" /keys || {
-        echo 'Retrying without overriding UID'
-        # Above line doesn't work on macOS + VirtualBox Docker
-        docker run --rm -it -v "$OUTPUT_DIR":/keys casperlabs/key-generator:"$TAG" /keys
-    }
-else
-    docker run --rm -v "$OUTPUT_DIR":/keys casperlabs/key-generator:DRONE-"$DRONE_BUILD_NUMBER" /keys
-fi
-
-
+TAG=${CL_VERSION:-"latest"}
+docker pull casperlabs/key-generator:"$TAG" &> /dev/null
+docker run --rm -it --user $(id -u):$(id -g) -v "$OUTPUT_DIR":/keys casperlabs/key-generator:"$TAG" /keys || {
+    echo 'Retrying without overriding UID'
+    # Above line doesn't work on macOS + VirtualBox Docker
+    docker run --rm -it -v "$OUTPUT_DIR":/keys casperlabs/key-generator:"$TAG" /keys
+}
 if [[ "$1" == "--test" ]]; then
     if [[ -f "$OUTPUT_DIR/node-id" ]] && \
        [[ -f "$OUTPUT_DIR/validator-pk" ]] && \
