--- manual-test/required-contracts.txt
@@ -1,3 +1,0 @@
-faucet
-test-payment-stored
-transfer-to-account-u512-stored
