--- manual-test/teardown.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-
-# Shutdown highway network
-cd ../../CasperLabs/hack/docker || exit
-make clean
