+++ network-tests/teardown.sh
@@ -1,0 +1,9 @@
+#!/usr/bin/env bash
+
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+
+cd "$DIR" || exit 1
+
+# Shutdown highway network
+cd ../hack/docker || exit 1
+make clean
