+++ network-tests/tests/common.py
@@ -6,16 +6,11 @@ from casperlabs_client import CasperLabsClient
 from casperlabs_client.abi import ABI
 
 THIS_DIRECTORY = Path(os.path.dirname(os.path.realpath(__file__)))
-CASPERLABS_ROOT_DIRECTORY = THIS_DIRECTORY.parent.parent.parent / "CasperLabs"
+NETWORK_TESTS_ROOT = THIS_DIRECTORY.parent
+PROJECT_DIR = THIS_DIRECTORY.parent.parent
 
-WASM_DIRECTORY = (
-    CASPERLABS_ROOT_DIRECTORY
-    / "execution-engine"
-    / "target"
-    / "wasm32-unknown-unknown"
-    / "release"
-)
-HACK_DOCKER_DIRECTORY = CASPERLABS_ROOT_DIRECTORY / "hack" / "docker"
+WASM_DIRECTORY = NETWORK_TESTS_ROOT / "wasm"
+HACK_DOCKER_DIRECTORY = PROJECT_DIR / "hack" / "docker"
 
 KEYS_DIRECTORY = HACK_DOCKER_DIRECTORY / "keys"
 FAUCET_PRIVATE_KEY_PEM_PATH = KEYS_DIRECTORY / "faucet-account" / "account-private.pem"
@@ -41,7 +36,7 @@ def faucet_fund_account(casperlabs_client, account_hash_hex, amount=10000000000)
     faucet_wasm_path = WASM_DIRECTORY / "faucet.wasm"
     assert faucet_wasm_path.exists(), (
         f"Needed wasm file: {faucet_wasm_path} does not exist.\n"
-        "It should be built in `build_contracts.sh`."
+        "It should be built in `build_contracts_in_buildenv.sh`."
     )
     session_args = ABI.args(
         [ABI.account("target", account_hash_hex), ABI.big_int("amount", amount)]
@@ -51,7 +46,7 @@ def faucet_fund_account(casperlabs_client, account_hash_hex, amount=10000000000)
         private_key=FAUCET_PRIVATE_KEY_PEM_PATH,
         session=faucet_wasm_path,
         session_args=session_args,
-        payment_amount=1000000,
+        payment_amount=10000000,
     )
     result = casperlabs_client.show_deploy(deploy_hash, wait_for_processed=True)
     assert (
